requirements.txt
streamlit_code_editor
